Meets design requirements!
I targeted CM, since it was the parameter that was farthest from the
constraint. I decreased CM by “flattening” the Cp distribution of the
bottom surface, so that the lift on the front of the airfoil was
increased and the lift on the aft of the airfoil was decreased. I used
the MDES design routine in XFoil to make these adjustments. The
resulting airfoil had a slightly thinner maximum thickness and camber,
but meets all design specifications.